Refactor the administrator/ structure
