--- resources/js/dashboard/pages/administrator/users/components/List.vue
@@ -1,26 +1,0 @@
-<template>
-  <div>
-    <b-table striped hover :items="items"></b-table>
-  </div>
-</template>
-
-<script>
-  export default {
-    data() {
-      return {
-        items: [
-          { age: 40, first_name: 'Dickerson', last_name: 'Macdonald' },
-          { age: 21, first_name: 'Larsen', last_name: 'Shaw' },
-          { age: 89, first_name: 'Geneva', last_name: 'Wilson' },
-          { age: 38, first_name: 'Jami', last_name: 'Carney' },
-          { age: 89, first_name: 'Geneva', last_name: 'Wilson' },
-          { age: 38, first_name: 'Jami', last_name: 'Carney' },
-          { age: 38, first_name: 'Jami', last_name: 'Carney' },
-          { age: 38, first_name: 'Jami', last_name: 'Carney' },
-          { age: 38, first_name: 'Jami', last_name: 'Carney' },
-          { age: 38, first_name: 'Jami', last_name: 'Carney' }
-        ]
-      }
-    }
-  }
-</script>
