+++ resources/js/dashboard/pages/administrator/users/components/ListUsers.vue
@@ -1,0 +1,16 @@
+<template>
+  <div>
+    <b-table striped hover :items="users"></b-table>
+  </div>
+</template>
+
+<script>
+  export default {
+    props: {
+      users: {
+        type: Array,
+        required: true
+      }
+    },
+  }
+</script>
